feat: small changes to privacy in stacks, added tests
