+++ treiber/src/test/kotlin/lincheck/EliminationStackTest.kt
@@ -1,11 +1,14 @@
-import org.jetbrains.kotlinx.lincheck.annotations.*
-import org.jetbrains.kotlinx.lincheck.strategy.stress.*
-import org.jetbrains.kotlinx.lincheck.check
-import org.jetbrains.kotlinx.lincheck.strategy.managed.modelchecking.*
-import org.junit.jupiter.api.*
+package lincheck
 
-class TreiberStackTest {
-    private val stack = TreiberStack<Int>()
+import elimination.EliminationStack
+import org.jetbrains.kotlinx.lincheck.annotations.Operation
+import org.jetbrains.kotlinx.lincheck.check
+import org.jetbrains.kotlinx.lincheck.strategy.managed.modelchecking.ModelCheckingOptions
+import org.jetbrains.kotlinx.lincheck.strategy.stress.StressOptions
+import org.junit.jupiter.api.Test
+
+class EliminationStackTest {
+    private val stack = EliminationStack<Int>(100)
 
     @Operation
     fun pop() = stack.pop()
