+++ treiber/src/test/kotlin/lincheck/TreiberStackTest.kt
@@ -1,12 +1,14 @@
-import elimination.EliminationStack
-import org.jetbrains.kotlinx.lincheck.annotations.Operation
-import org.jetbrains.kotlinx.lincheck.check
-import org.jetbrains.kotlinx.lincheck.strategy.managed.modelchecking.ModelCheckingOptions
-import org.jetbrains.kotlinx.lincheck.strategy.stress.StressOptions
-import org.junit.jupiter.api.Test
+package lincheck
 
-class EliminationStackTest {
-    private val stack = EliminationStack<Int>(100)
+import TreiberStack
+import org.jetbrains.kotlinx.lincheck.annotations.*
+import org.jetbrains.kotlinx.lincheck.strategy.stress.*
+import org.jetbrains.kotlinx.lincheck.check
+import org.jetbrains.kotlinx.lincheck.strategy.managed.modelchecking.*
+import org.junit.jupiter.api.*
+
+class TreiberStackTest {
+    private val stack = TreiberStack<Int>()
 
     @Operation
     fun pop() = stack.pop()
